Add files via upload
This script file will introduce the notion of pipelines and how scikit-learn allows for transformers and estimators to be chained together and used as a single unit. Pre-processing techniques will be then  introduced as a way to enhance model performance and pipelines will be the glue that ties together concepts in the prior chapters.